Fix invalid id WWW-Authenticate behavior conflict with common .htaccess pattern
**Why:** It is common for .htaccess files - including the default one
provided by Drupal Core - to coerce the `Authorization` header to the
empty string if not specified by the client. Without this fix, that
results in a `WWW-Authenticate: Basic` header response which causes
the browser to show a useless basic authentication dialog.

For more info see https://github.com/thephpleague/oauth2-server/issues/1162